+++ main/http_server/handler_system.h
@@ -1,0 +1,4 @@
+#pragma once
+
+esp_err_t GET_system_info(httpd_req_t *req);
+esp_err_t PATCH_update_settings(httpd_req_t *req);
